// apps/registry/codemods/eslint/biome/migrate-rules/types/biome.d.ts
export type PlainIndentStyle = "tab" | "space";
export type LineEnding = "lf" | "crlf" | "cr";
/**
 * Validated value for the `line_width` formatter options
 *
 * The allowed range of values is 1..=320
 */
export type LineWidth = number;
export type QuoteStyle = "double" | "single";
export type StringSet = string[];
export type ArrowParentheses = "always" | "asNeeded";
export type QuoteProperties = "asNeeded" | "preserve";
export type Semicolons = "always" | "asNeeded";
/**
 * Print trailing commas wherever possible in multi-line comma-separated syntactic structures.
 */
export type TrailingComma = "all" | "es5" | "none";
export type RuleConfiguration = RulePlainConfiguration | RuleWithOptions;
// ...
export type Overrides = OverridePattern[];
export type VcsClientKind = "git";

/**
 * The configuration that is contained inside the file `biome.json`
 */
export interface Configuration {
  /**
   * A field for the [JSON schema](https://json-schema.org/) specification
   */
  $schema?: string | null;
  /**
   * Specific configuration for the Css language
   */
  css?: CssConfiguration | null;
  /**
   * A list of paths to other JSON files, used to extends the current configuration.
   */
  extends?: StringSet | null;
  /**
   * The configuration of the filesystem
   */
  files?: FilesConfiguration | null;
  /**
   * The configuration of the formatter
   */
  formatter?: FormatterConfiguration | null;
  /**
   * Specific configuration for the JavaScript language
   */
  javascript?: JavascriptConfiguration | null;
  /**
   * Specific configuration for the Json language
   */
  json?: JsonConfiguration | null;
  /**
   * The configuration for the linter
   */
  linter?: LinterConfiguration | null;
  /**
   * The configuration of the import sorting
   */
  organizeImports?: OrganizeImports | null;
  /**
   * A list of granular patterns that should be applied only to a sub set of files
   */
  overrides?: Overrides | null;
  /**
   * The configuration of the VCS integration
   */
  vcs?: VcsConfiguration | null;
}
/**
 * Options applied to CSS files
 */
export interface CssConfiguration {
  /**
   * Formatting options
   */
  formatter?: CssFormatter | null;
  /**
   * Parsing options
   */
  parser?: CssParser | null;
}
// ...
export interface CssParser {
  /**
   * Allow comments to appear on incorrect lines in `.css` files
   */
  allowWrongLineComments?: boolean | null;
}
/**
 * The configuration of the filesystem
 */
export interface FilesConfiguration {
  /**
   * A list of Unix shell style patterns. Biome will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * Tells Biome to not emit diagnostics when handling files that doesn't know
   */
  ignoreUnknown?: boolean | null;
  /**
   * A list of Unix shell style patterns. Biome will handle only those files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * The maximum allowed size for source code files in bytes. Files above this limit will be ignored for performance reasons. Defaults to 1 MiB
   */
  maxSize?: number | null;
}
/**
 * Generic options applied to all files
 */
export interface FormatterConfiguration {
  enabled?: boolean | null;
  /**
   * Stores whether formatting should be allowed to proceed if a given file has syntax errors
   */
  formatWithErrors?: boolean | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * The size of the indentation, 2 by default (deprecated, use `indent-width`)
   */
  indentSize?: number | null;
  /**
   * The indent style.
   */
  indentStyle?: PlainIndentStyle | null;
  /**
   * The size of the indentation, 2 by default
   */
  indentWidth?: number | null;
  /**
   * The type of line ending.
   */
  lineEnding?: LineEnding | null;
  /**
   * What's the max width of a line. Defaults to 80.
   */
  lineWidth?: LineWidth | null;
}
/**
 * A set of options applied to the JavaScript files
 */
export interface JavascriptConfiguration {
  /**
   * Formatting options
   */
  formatter?: JavascriptFormatter | null;
  /**
   * A list of global bindings that should be ignored by the analyzers
   *
   * If defined here, they should not emit diagnostics.
   */
  globals?: StringSet | null;
  organize_imports?: JavascriptOrganizeImports | null;
  /**
   * Parsing options
   */
  parser?: JavascriptParser | null;
}
// ...
export interface JavascriptOrganizeImports {}
/**
 * Options that changes how the JavaScript parser behaves
 */
export interface JavascriptParser {
  /**
   * It enables the experimental and unsafe parsing of parameter decorators
   *
   * These decorators belong to an old proposal, and they are subject to change.
   */
  unsafeParameterDecoratorsEnabled?: boolean | null;
}
/**
 * Options applied to JSON files
 */
export interface JsonConfiguration {
  /**
   * Formatting options
   */
  formatter?: JsonFormatter | null;
  /**
   * Parsing options
   */
  parser?: JsonParser | null;
}
// ...
export interface LinterConfiguration {
  /**
   * if `false`, it disables the feature and the linter won't be executed. `true` by default
   */
  enabled?: boolean | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
  /**
   * List of rules
   */
  rules?: Rules | null;
}
export interface Rules {
  a11y?: A11Y | null;
  /**
   * It enables ALL rules. The rules that belong to `nursery` won't be enabled.
   */
  all?: boolean | null;
  complexity?: Complexity | null;
  correctness?: Correctness | null;
  nursery?: Nursery | null;
  performance?: Performance | null;
  /**
   * It enables the lint rules recommended by Biome. `true` by default.
   */
  recommended?: boolean | null;
  security?: Security | null;
  style?: Style | null;
  suspicious?: Suspicious | null;
}
/**
 * A list of rules that belong to this group
 */
export interface A11Y {
  /**
   * It enables ALL rules for this group.
   */
  all?: boolean | null;
  /**
   * Enforce that the accessKey attribute is not used on any HTML element.
   */
  noAccessKey?: RuleConfiguration | null;
  /**
   * Enforce that aria-hidden="true" is not set on focusable elements.
   */
  noAriaHiddenOnFocusable?: RuleConfiguration | null;
  /**
   * Enforce that elements that do not support ARIA roles, states, and properties do not have those attributes.
   */
  noAriaUnsupportedElements?: RuleConfiguration | null;
  /**
   * Enforce that autoFocus prop is not used on elements.
   */
  noAutofocus?: RuleConfiguration | null;
  /**
   * Disallow target="_blank" attribute without rel="noreferrer"
   */
  noBlankTarget?: RuleConfiguration | null;
  /**
   * Enforces that no distracting elements are used.
   */
  noDistractingElements?: RuleConfiguration | null;
  /**
   * The scope prop should be used only on <th> elements.
   */
  noHeaderScope?: RuleConfiguration | null;
  /**
   * Enforce that non-interactive ARIA roles are not assigned to interactive HTML elements.
   */
  noInteractiveElementToNoninteractiveRole?: RuleConfiguration | null;
  /**
   * Enforce that interactive ARIA roles are not assigned to non-interactive HTML elements.
   */
  noNoninteractiveElementToInteractiveRole?: RuleConfiguration | null;
  /**
   * Enforce that tabIndex is not assigned to non-interactive HTML elements.
   */
  noNoninteractiveTabindex?: RuleConfiguration | null;
  /**
   * Prevent the usage of positive integers on tabIndex property
   */
  noPositiveTabindex?: RuleConfiguration | null;
  /**
   * Enforce img alt prop does not contain the word "image", "picture", or "photo".
   */
  noRedundantAlt?: RuleConfiguration | null;
  /**
   * Enforce explicit role property is not the same as implicit/default role property on an element.
   */
  noRedundantRoles?: RuleConfiguration | null;
  /**
   * Enforces the usage of the title element for the svg element.
   */
  noSvgWithoutTitle?: RuleConfiguration | null;
  /**
   * It enables the recommended rules for this group
   */
  recommended?: boolean | null;
  /**
   * Enforce that all elements that require alternative text have meaningful information to relay back to the end user.
   */
  useAltText?: RuleConfiguration | null;
  /**
   * Enforce that anchors have content and that the content is accessible to screen readers.
   */
  useAnchorContent?: RuleConfiguration | null;
  /**
   * Enforce that tabIndex is assigned to non-interactive HTML elements with aria-activedescendant.
   */
  useAriaActivedescendantWithTabindex?: RuleConfiguration | null;
  /**
   * Enforce that elements with ARIA roles must have all required ARIA attributes for that role.
   */
  useAriaPropsForRole?: RuleConfiguration | null;
  /**
   * Enforces the usage of the attribute type for the element button
   */
  useButtonType?: RuleConfiguration | null;
  /**
   * Enforce that heading elements (h1, h2, etc.) have content and that the content is accessible to screen readers. Accessible means that it is not hidden using the aria-hidden prop.
   */
  useHeadingContent?: RuleConfiguration | null;
  /**
   * Enforce that html element has lang attribute.
   */
  useHtmlLang?: RuleConfiguration | null;
  /**
   * Enforces the usage of the attribute title for the element iframe.
   */
  useIframeTitle?: RuleConfiguration | null;
  /**
   * Enforce onClick is accompanied by at least one of the following: onKeyUp, onKeyDown, onKeyPress.
   */
  useKeyWithClickEvents?: RuleConfiguration | null;
  /**
   * Enforce onMouseOver / onMouseOut are accompanied by onFocus / onBlur.
   */
  useKeyWithMouseEvents?: RuleConfiguration | null;
  /**
   * Enforces that audio and video elements must have a track for captions.
   */
  useMediaCaption?: RuleConfiguration | null;
  /**
   * Enforce that all anchors are valid, and they are navigable elements.
   */
  useValidAnchor?: RuleConfiguration | null;
  /**
   * Ensures that ARIA properties aria-* are all valid.
   */
  useValidAriaProps?: RuleConfiguration | null;
  /**
   * Elements with ARIA roles must use a valid, non-abstract ARIA role.
   */
  useValidAriaRole?: RuleConfiguration | null;
  /**
   * Enforce that ARIA state and property values are valid.
   */
  useValidAriaValues?: RuleConfiguration | null;
  /**
   * Ensure that the attribute passed to the lang attribute is a correct ISO language and/or country.
   */
  useValidLang?: RuleConfiguration | null;
  [k: string]: unknown;
}
// ...
export interface Performance {
  /**
   * It enables ALL rules for this group.
   */
  all?: boolean | null;
  /**
   * Disallow the use of spread (...) syntax on accumulators.
   */
  noAccumulatingSpread?: RuleConfiguration | null;
  /**
   * Disallow the use of the delete operator.
   */
  noDelete?: RuleConfiguration | null;
  /**
   * It enables the recommended rules for this group
   */
  recommended?: boolean | null;
  [k: string]: unknown;
}
/**
 * A list of rules that belong to this group
 */
export interface Security {
  /**
   * It enables ALL rules for this group.
   */
  all?: boolean | null;
  /**
   * Prevent the usage of dangerous JSX props
   */
  noDangerouslySetInnerHtml?: RuleConfiguration | null;
  /**
   * Report when a DOM element or a component uses both children and dangerouslySetInnerHTML prop.
   */
  noDangerouslySetInnerHtmlWithChildren?: RuleConfiguration | null;
  /**
   * It enables the recommended rules for this group
   */
  recommended?: boolean | null;
  [k: string]: unknown;
}
// ...
export interface OrganizeImports {
  /**
   * Enables the organization of imports
   */
  enabled?: boolean | null;
  /**
   * A list of Unix shell style patterns. The formatter will ignore files/folders that will match these patterns.
   */
  ignore?: StringSet | null;
  /**
   * A list of Unix shell style patterns. The formatter will include files/folders that will match these patterns.
   */
  include?: StringSet | null;
}
// ...
export interface VcsConfiguration {
  /**
   * The kind of client.
   */
  clientKind?: VcsClientKind | null;
  /**
   * The main branch of the project
   */
  defaultBranch?: string | null;
  /**
   * Whether Biome should integrate itself with the VCS client
   */
  enabled?: boolean | null;
  /**
   * The folder where Biome should check for VCS files. By default, Biome will use the same folder where `biome.json` was found.
   *
   * If Biome can't find the configuration, it will attempt to use the current working directory. If no current working directory can't be found, Biome won't use the VCS integration, and a diagnostic will be emitted
   */
  root?: string | null;
  /**
   * Whether Biome should use the VCS ignore file. When [true], Biome will ignore the files specified in the ignore file.
   */
  useIgnoreFile?: boolean | null;
}